scan_api_internal/pyproject.toml
[tool.poetry]
name = "fastapi-example"
version = "0.1.0"
description = ""
authors = ["Jun Fujima <jf.devvv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
cryptography = "*"
fastapi = "*"
fastapi-limiter = "*"
fastapi-pagination = {version = "==0.8.1", extras = ["all"]}
joblib = "*"
networkx = "*"
numpy = "*"
pandas = "*"
pydantic = "*"
pymysql = "*"
scipy = "*"
sqlalchemy = "*"
tqdm = "*"
ulid-py = "*"
uvicorn = "*"
python-dotenv = "^0.19.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
autoflake = "*"
black = "==21.7b0"
isort = "*"
